criando .h do processo
